Moved common logic to extension method.
